requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asgiref==3.8.1
attrs==25.3.0
authentication==1.1.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.6.1
Flask==3.1.1
frozenlist==1.7.0
googletrans==4.0.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langdetect==1.0.9
MarkupSafe==3.0.2
multidict==6.6.3
openai==0.28.0
passlib==1.7.4
propcache==0.3.2
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rest-framework-simplejwt==0.0.2
rpds-py==0.25.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.20.1
